split rand.rs
